pyproject.toml
[project]
name = "visualization"
version = "0.1.0"
description = "Standalone FastAPI visualization for Oracle property graph types"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.129.0",
    "envyaml>=1.10.211231",
    "oci>=2.144.0",
    "oracledb>=2.3.0",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.1",
    "uvicorn>=0.41.0",
    "httpx>=0.28.0",
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
